requirements.txt
absl-py==0.9.0
appdirs==1.4.3
astroid==2.3.3
certifi==2019.9.11
chardet==3.0.4
Click==7.0
colorthief==0.2.1
colour==0.1.5
cPython==0.0.5
decorator==4.4.2
distlib==0.3.0
Django==2.2.7
djangorestframework==3.10.3
EasyProcess==0.2.10
entrypoint2==0.2
filelock==3.0.12
Flask==1.1.1
idna==2.9
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
Pillow==7.0.0
pylint==2.4.4
pymongo==3.9.0
pyscreenshot==1.0
pytz==2019.3
requests==2.23.0
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
virtualenv==20.0.13
webcolors==1.3
Werkzeug==0.16.0
wrapt==1.11.2
